samples: knot: Manage pipes using fifos
Avoid pipe overflow using fifos. Only one message is put in pipe at
time.
Since neither NET nor PROTO thread know KNoT protocol, it is
transparent to both of them which was the message. Whenever any of
them receive a message, it is queued and sent through the pipe when
it is free.

Signed-off-by: Paulo Serra Filho <ptvsf@cesar.org.br>